requirements.txt
langchain==0.3.27
langchain-community==0.3.30
langchain-core==0.3.78
langchain-openai==0.3.35
langchain-text-splitters==0.3.11
python-dotenv==1.1.1
ollama==0.6.0
tavily-python == 0.7.12
